Canvas Drawing and Server Data Transfer
Canvas can be drawn on! Also features a clear button to get rid of those pesky mistakes.

Submit button will send the data to the server where it will calculate the percentage of the canvas that consists of black pixels.

Nerual Network Libary synaptic is loaded in, but yet to be used.